This ups the minor revision number to 0.3.0.  Worthwhile because now I have BOTH an erfc AND a Q evaluation of p-value.  I could certainly prettify sts_monobit, but since I generally think that it isn't that great a test (although it does indicate how starkly many rng's FAIL to be even this random) I won't do so right away.
Next (after tagging and resync'ing) is going to be adding more tests.
At this point adding a test should be pretty easy, given the hopefully
reusable routines I have written to do the pre- and post- processing.
All I really have to do is input the expected values, write a loop to
generate the "experimental" statistic, and pass everything on to a
standard set of tools for outputting the results and deciding on the
quality of the results.