OK, we've learned the hard way that some bits in e.g. boroshi don't change AT ALL, EVER.  Which makes it pretty hard to be random, of course.
So we're going to invent a new tool -- rgb_persist(), which doesn't
(yet) to a formal statistical test.  It just is going to dump successive
unsigned ints from the rng (bitwise) AND maybe run a string of &'s on
the string of ints returned.  If they share any 1 bits, the successive
&'s will preserve them a LOT longer than permitted by binary flips on
the slots.

This could be made into a fairly powerful bitlevel sequential
correlation test in several ways.  We'll investigate them as we go, but
one reason to write this now is that I'm not quite convinced that what
I'm seeing isn't some sort of bug in the get_bit() routine or the like.