OK, added a few minor changes to manage the bits issue yet another way.
Really, I'm going to have to figure out a consistent way of indicating
whether a test can have size OR bits OR both OR neither specified.

Also, it would be really lovely to have another outer loop and to
present the lowest p in a set of (say) ten runs of a test combo.
Although in many cases running with -s 10x larger should do the same
thing, really.