This is actually a fairly functional diehard test!
I think that we can actually implement a test for the uniformity of
p-values as suggested by NIST to run on TOP of the existing confidence
interval test.  This would actually break the p-distribution down by
interval and return a p-value of its own computed against the assumption
of uniformity.  Or I could get fancier and try kolmogorov-smirnov, if
GSL doesn't have one and I work hard enough to program one.  If this is
really distinct -- it isn't clear that it is.