This is still fairly screwed up, at least in the sense that it doesn't look like rgb_bitdist works.  Curiously, it LOOKS like it WORKS -- walking through the code, it looks VERY much like it is collecting two bits at a time and correctly incrementing the correct bit count in the correct vector.
The final histogram, however, comes out wrong, wrong, wrong.

I may have to make this simpler.  Or maybe I'm doing something else
wrong -- come to think of it, the totals in the histograms shouldn't
equal the number of samples for EACH value of the ntuple, only the
total should sum to the number of samples.  Maybe this is what is
wrong...