This is now working.  Working amazingly well, actually.  Well enough to double the size of the bits in rgb_bitdist_test().
The one MAJOR remaining problem is that I cannot use samples for tests
that return a vector of pvalues.  Oh, and it is fairly difficult to
pass arguments to the testing function in when it is an argument TO
samples().

This means that I have two itty bitty problems to solve -- one is to
pass in parameters (possibly by making them global variables).  This
makes sense IF I want to be able to control them from the command line
anyway.  The other is to return a vector of pvalues.  The only way I
can think of doing this is to make pvalues[] a global vector as well
of length (say) 1K.  This puts an upper bound on the number of pvalues
that can be returned by a test, but that SHOULDN'T be much of a problem,
as it is really a question of what granularity one wishes to evaluate p
at.

Anyway, just a BIT more work and rgb_bitdist should be production ready,
AND I should be perfectly ready to clean up p-sampling and testing as
separate entities in the other tests.