This is a fairly major fix -- I was truncating blen in bits.c at the sizeof(uint), not 8*sizeof(uint).  One lesson is that this truncation isn't right anyway.  We rather need to just punt/die.
I'm wondering now if the apparent failure that is still present
(although not nearly as bad) for the larger ntuples is because fewer
bins pass the cutoff in the formation of the primary sample pvalue(s).

We might just try lowering this cutoff a bit.  I don't know exactly what
a "degree of freedom" is, but we do need to be pretty careful with it.