This is the last checkin of rand_rate AS rand_rate.  I'm going to change the name of this suite to dieharder.  I'm also going to change the test numbering schema and option naming so that e.g.
  -d diehard test number
  -r rgb test number
  -s sts test number

where -1 runs all tests of a given kind, 0 lists a description of all
tests in the suite, -2 runs all tests of a given kind EXACTLY as they
are run in the original code, if possible -- I'm not sure I'm going to
test overlapping bit strings drawn from a single int just to bump the
count of "random numbers" unless the test explicitly calls for it and it
makes sense, as there is this thing about each sample being
"independent" that worries me with overlapping draws.

  -p number of p-values in final KS (and possibly other) test(s).  This
is the number of times each "test" is run with independent random number
seeds (as the DEFAULT from now on).  This defaults to 100, which is
actually a lot and very reasonable but which can be increased if one is
in doubt about whether the distribution of p returned by "the test" is
in fact uniform.

  -t number of test samples that go into making up a single test.  This
is NOT always a free parameter -- in many of the diehard tests
especially, the number of e.g. points drawn from a 2d or 3d volume in
the minimum distance tests is fixed, and varying it would vary the
target distribution and test statistic.  Although this is a bit
unfortunate, since varying the number of test samples is an excellent
way to make marginal failures in the distribution of p resolve into
clear failures, we either live with it or derive general forms for the
asymptotic target distributions as a function of the number of samples
or do simulations and empirically deduce forms ditto (as Marsaglia and
others appear to have done).  For the moment we'll live with it.

  -b bitstring width.  Some tests are applied only to samples that are
"bitstrings" (as opposed to e.g. lists of unsigned integers) of
user-specifiable length.  One reason to limit the tests in this way is
to avoid numerical difficulties in e.g. evaluating P_binomial(k,p,n),
where one can easily under or overflow and end up with garbage or a gsl
fault.  Another is to "free" some of the existing tests that have a very
specific size of bitstring that they look at so that this can be varied
when the target distribution can still be computed as a function of
bitstring size.  This will be overridden as necessary, like -t, for
tests that really do require a fixed bitstring size to approach the
known target distribution.

  -n ntuple or window width.  A number of tests look at bit ntuples.  An
ntuple is a set of n consecutive bits drawn from a bitstring (possibly
of -b specified width) or vector of random uints (possibly of -t
specified length).  ntuples are always drawn relative to a bit offset
specified from the right (least significant) with 0 being the rightmost
bit, with cyclic boundary conditions on the entire bitstring >>or<<
sample uint vector, so drawing an ntuple cannot fail for any offset
within the number of significant bits returned by the generator (which
MAY NOT BE 32, or even 31 -- some generators return as few as 16
significant bits!).

For example, an 8-bit bitstring might be:

  01100111

and all the 3-tuples drawn from it are given by

  offset    3-tuple
=====================
    0        111
    1        011
    2        001
    3        100
    4        110
    5        011
    6        101  <- Note wrap around!
    7        110  <- Ditto!

A general purpose

  get_bit_ntuple(*bitstring,bitstring size,ntuple size,offset)

routine is provided that is used by many tests to get ntuples from a
uint *bitstring of given >>uint vector length<< (not bitstring length).

Other test controls may be added as well, but these are what I'm going
to document right now.  Mostly I'm checking in all the placeholders
required for the rest of the diehard tests so I can start to knock them
off systematically.  Sure hope I'm past major rewrites!