I'm most of the way through adding support for
  dieharder -o testrands.dat

so that I can turn around and run

  dieharder -f testrands.dat -r 2

to run rgb test 2 on file data with the default parameters.  -o will
supercede any other commands (in work.c) and get executed FIRST, but one
CAN use

  dieharder -o testrands.dat -r 2

to simultaneously create an output file and use it as input for a test
or suite of tests.  I think -- this may take some tuning to get just
right...