This is the bitstream test.  I'm down to operm5, which looks relatively "difficult", and two count the ones tests, which look easy.  I'll probably do easy first, then finish off with operm5.
I'm not absolutely certain that bitstream is valid -- BOTH mt1997 and
rndldx2 fail it!  Of course, I've implemented it to be more sensitive
than diehard was with a much more serious and revealing KS test on the
larger number of psamples.  I'm going to try the semi-hardware rng and
then the hardware rng.  If those fail it, the test is probably flawed
in some subtle way...