TAG 2.24.2
I fixed the -q(uiet) flag so that it works, and decided to bundle all of
the above bugfixes and feature shifts into a new minor minor bump.  This
is the "official" GBT build, if Dirk concurs when he tries it.

As far as I know, we're at least back where we were when we started
implementing GBT and then some.

Now, if only I could figure out the trunk thing.  Maybe I'll try doing a
checkout and build without the trunk movement...