This is pretty much a fully functional version.  I fixed the last bug, which turned out to be that I had the right idea but missed the:
  I + P + P* + P^2 + P*^2 + ...

and was accidentally doing (when I empirically noted that including k=0
in the sum was necessary) either:

  I + P + P^2...

or

  2I + P + P* + ...

Added a conditional to eliminate the second k = 0 instance in the core
loop summing the fi*fj and got perfect correspondance through -n 3.  The
-n 3 matrix is different, but pretty obviously a permutation of the
AKM matrix as to be expected since the permutation order is arbitrary
and they probably didn't use lexical ordering.  The eigenvalues,
however, are dead on the same!

So I'm back up to the point where I should be able to form a test
statistic somehow.  I have to covariance matrix precisely in hand --
something at this point about weak inverses (weak because it is singular
and has no inverse?).  A bit curious, actually.  We'll see.

  rgb