This might be ready for a production snapshot, or a testing snapshot, or something like that.  At any rate, I have BOTH fixed the generator problem (where rngs have numbers that can change as new generators are added.  I fixed this by significantly increasing the size of the types[] vector and working moderately hard to pack it with generators sparsely and sanely -- sparse so there is plenty of room, sane because the space is now blocked out by general category with room to grown.
This will NOT GUARANTEE that the GSL will not change its internal order,
and dieharder will continue to follow the GSL order where they overlap
as nothing else makes sense.

I >>also<< think I've fixed -r 5, rgb_permutations.  This is a
non-overlapping permutations test and runs to more or less arbitrary
order (and seems to work quite nicely where it runs, although it gets
quite slow when we get to very large permutations as one might expect).

I think it is time to bump the minor-minor number to 11 and post it, for
the time being with -r 5 still marked out as experimental in the
help listing.