Once again, this is a transient checkin so that I can move and rearrange.  Things are actually coming together nicely (Dirk), and my remaining couple of chores are to make random number selection and test selection work much the same way, with a single three-digit index into a big vector of tests and a very simple name matcher/parser so that tests can be selected by number or by name.  I've built and tested the parser logic itself in parsecl.c, so it really should be no trouble to finish it, once I've got a bit of renaming out of the way.
The separation of "dieharder" CLI operation and Rdh UI operation should
be a piece of cake when I'm done -- a tiny handful of ifdefs, a single
place you'll need "Rdh" code to set up the basic call in place of
parsecl.c, and basically two commands that will be in your main
interpreter loop -- one to pick and initialize an rng, one to invoke a
test.  Output is all done in output.c, all dieharder output is in
ifdefs, with a very few task-specific CLI pathways to e.g. print help or
cause a list of rands to be dumped to stdout.  dieharder output is
at this point completely controllable via a set of additive flags as we
discussed -- you can truly roll your own output, trivially, including
your own favorite field separator, while the default table-style output
is still perfectly readable.

I'm still holding off on a full checkin until I get this all finished,
because it will very definitely need some serious testing and there
isn't anything to test yet -- some things are done and work perfectly,
but some things have yet to be done and are therefore "broken".

Just FYI -- I was hoping to finish today, but might not make it unless I
give up any hope of sleep.  I've got around 3-4 hours more work to the
proper 3.28.0beta release.  More if you include fixing the documentation
to match (but I may do that while you are playing with the beta).