This records a validation script to use with aespipe to produce a "standard run" of dieharder in -v3.  aespipe with the fixed, trivial 256-bit key in aeskey below, is used to encrypt /dev/zero into a stdin stream and fed to dieharder -a.  The encrypted stream should be as close to "truly random" as we can currently manage with simple, reasonably fast tools.  The interesting thing is that this stream actually PASSES ALL OF THE TESTS in dieharder, even the "known bad" tests such as diehard_operm5.  This makes it very, very useful for comparison purposes.  For example, for the first time ever, I feel like I can now say that mt19937 actually FAILS dieharder (or has weaknesses that are explicitly exposed by dieharder) when it consistently has tests (even very specific tests for very specific length ntuples) on which it is weak or fails or exhibits high bias in its output pvalues.
To be fair, passing all of the tests isn't necessarily a good thing,
since there are over 100 of them including ntuples.  One expects 1/100
or 1/200 or thereabouts to be weak for a PERFECT RNG on most runs.
Eyeballing the distribution of final P in the aespipe run reveals that
dieharder still produces a weak high bias in the final distribution of
pvalues, but this is very much in line for the bias revealed by
rgb_kstest_test and is therefore very likely an artifact of using -p 100
as the default for most of the tests in -a(ll).

I'm going to run the validation line:

  cat /dev/zero | aespipe -P aeskey | ./dieharder -g 200 -a

with -p 1000 just for grins (which will take the rest of the day, I
expect) and see if it doesn't push the final distribution right back
where it belongs, with less visible bias towards the 0.9-1.0 range and
away from 0.0-0.1 on the bottom.

Still, a perfect PASS for a nearly perfect generator.  How cool is that?