This is most of the threefish stuff required, but I'm still having trouble with the big/smallendian conversions apparently needed by skein in threefish.  One function that is supposed to be defined automagically is coming out UNdefined in the linker, which is "bad".  I may have to ask David Bauer how he got this to compile.  Note that I've added both bauer and emelko's current round of bug reports and remarks to the Bugs directory below.  David in particular has been really looking hard at kstests, and with good justification.  The kstest is apparently very poorly defined even in stats texts and the literature.  It is apparently more broken in R than it is in dieharder, and it is still a bit broken in dieharder.
As is so often the case in dieharder problems, pushing the test suite to
new limits exposes weaknesses in code that has long been taken for
granted because it has never been used for a rigorous analysis of this
sort.  But it NEEDS a precise ks or kuiper test, not just a sorta-useful
approximate one, or one cannot rely on its statements of weakness or
failure.

Anyway, this checkin is still broken but is within one #define or so of
working, I think, once I figure out how to do it without violating the
code in the brg_endian.h include that is supposed to automagically
select the right Skein function that is currently undefined.