FINALLY I got threefish to work.  brg_endian.h was broken as shit; it starts off by pulling something from crypt.h that is obviously broken on modern linux boxes (at least my Fedora 11 x86_64 box).  The remaining code looks quite general and seems to work, although I have to admit I absolutely hate crap code like this -- it smacks of aimk, imake, and other crap tools that detect platform type using some sort of transient trace from one tool or another that breaks three years later (or rather, requires yet another conditional).  I'll leave this in for now in case somebody wants to port to sparc or some bigendian platform, but since we are using threefish only to make random numbers and don't care to ever decrypt the stream of 0's or whatever it is applied to, I honestly doubt that it matters.  Getting endianness wrong sounds like it is at worst an extra byte shuffle.
Either way, this will be 3.29.2beta and I'll put it up on the dieharder
website in a few minutes (after a full -a run of threefish passes).  I
may add a comment to brg_endian.h indicating my hack, lest people be
tempted to use it as if it weren't modified.

Grrr.  I'm REALLY tempted to just strip it to the two line definition
that is all that matters in skein_port.h and screw the whole "automagic"
thing.  Robust code is robust code, and there are bound to be
intrinsically portable ways to handle endianness IF it is really
necessary in the first place.

At least this finally liberates me to move on and work on kstest and
kuiper again.  That's been on hold for a few days, but I'm feeling like
we're getting close to having one or the other work "perfectly" (if I
can find and add the missing O(1/N) correction terms from the
literature).