I'm checking in a lot of changes down below.  -m is implemented and documented.  -k is implemented and documented.  The man page is fixed (post good-kstest and aes/threefish).  The endian bug went away when I refreshed the include files, making me wonder if it wasn't some sort of strange GBT stuff and not a real problem -- I left in the endian code in configure.ac but don't use it.  I re-fixed diehard_runs.c -- it was broken post patching but now seems good.  I filed some documentation and bug reports.  I fixed a number of pernicious warnings about needing casts (one still remains in threefish, but it is David's and I don't know how to fix it).  I worked on dieharder.html.in pretty substantially to get it to match all of the above.
Next, -x (and maybe -X).

(p.s. -- Welcome David.)

(p.p.s. -- I'm still testing -- sigh, forever -- but it looks like all
non-deprecated tests are working OK in this snap, and that the -m
feature works nicely.  I documented timings for the k options, and
basically it comes down the kstest being too slow to do large numbers of
samples without switching over to the asymptotic form of the test at
some point.  I mean, going from three minute runs to over three hours
and still counting when I quit for a factor of 10 difference in the
number of samples, really serious nonlinear gains in the amount of
work/time required, and this was still -k 1 with Marsaglia's more modest
speedup, not even the "exact" mode with no speedup at all.  

This will quite possibly require some further hacking of the boundaries
for a crossover that is "practical" and not too inaccurate as we gain
experience with our own patience, especially as we implement a -x like
option that just keeps crankin' on the number of samples to hit a
prespecified tolerance for failure.)