This is a not-quite-yet-broken checkin of -Xtreme mode changes.  Three new control variables are in place.  They are parsed (untested).  They are used in std_test to allocate much larger pvalue vectors (Xcutoff in size) at test initialization time.  I'm JUST READY to hack into the main std_test execution loop with case switches or other conditionals and implement at least resolve ambiguity and ttd modes.  But as usual, I have to go in and teach.  At the moment, though, DH still builds and runs -a correctly, so it seems like a good idea to check in a still working snap in case I break everything and want to start over from here.
Oh, I also am cleaning up a bit and made the multiply_p variable (-m
option) a double, so you CAN enter -m 0.1 and run only 10 psamples for a
fast version of -a(ll).  At this point a lot of debugging is just
ensuring that all the tests run, and it is a PITA to wait 30+ minutes
for a -a(ll) run to get through.  So you now CAN test fewer than the
default number of psamples in an -all run, even though most people won't
use the option in actual testing.  The usual usage, -m 10 or -m 100,
works fine still.