Wow, a lot of stuff.  This checkin contains a working -Y 2 option for "test to destruction" where ttd is by default a return pvalue of 0.000001 or less OR getting to 10000 samples alive (both parameters can be set on the command line with -X tolerance and -Z cutoff).  I actually did it two ways, and will keep the second (better) one and shortly remove the cruft in std_test.c.  In addition I had to update the help, I updated the output routines in output_rnds so I could dump a list of formatted floats (to test another rng tester that alas was so broken it couldn't read any format I tried anyway), I fixed and updated the man page, I got rid of the old overlap variable (no longer desireable or necessary, although I have a bit of cruft left behind to clean up still).
As a result of the initial ttd test, I am certain that there is a
problem with diehard_dna, one that causes it to fail aes at 0.000001 in
1500-4000 samples.  This is odd, since this test has an "exactly"
computed mean and sigma target.  I may try threefish in a second to see
if it fails too, in the same general order.  Haven't done the auto-xor
generator thing yet.

I still have to implement -Y 1 (resolve ambiguity mode) where it will
force a test to fail or come back up with more samples, but it should be
straightforward.  However, it is almost 2 am and I teach way too early,
so it is off to bed.

    rgb