There are many changes herein.  I put the autoreconf command into autogen.sh; I'm trying to make this one script clean up all manner of GBT evil that might be leftover from a checkin on this machine or that machine.  I'll leave Dirk's exotically named script in place for the time being.  This leaves only the -I m4 weirdness reported by libtoolize, and I added a couple of lines to maybe start to fix that but I have to do something (I think) in one or more of the Makefile.am files to match and I haven't quite figured out what yet.
The big news of the checkin is Stephen Moenkeheus's fix for
diehard_operm5, which appears to work at least well enough to take
operm5 out to 5000 psamples (at which point I got bored).  It is no
longer a primary problem, in other words, whether or not it is
sufficiently exact to go to a gazillion psamples or a gazillion
tsamples.  I haven't played with it to see what kind of parametric
variability it has.  He did make it work with EITHER overlapping OR
non-overlapping samples, and it should more or less directly compare
to rgb_permutations in the latter invocation, something I will
eventually look at, but at the moment it suffices to note that
it empirically works (passes) applied to  AES either way.

Note that three new documents are in the doc directory -- a paper by
Tsang and Pang, Stephen's email to me detailing how he did the fix,
and a mathematica file that computed the permutation correlation matrix.
I wrote C code that computes the matrix in my abortive attempt to fix
this myself, and I SHOULD be able to check (at long last) if we get the
same matrix.  If we do, then I should be able to at VERY long last write
an rgb_opermutations that allows overlapping permutations out to perhaps
8 or 9 to be tested (limited by system speed and boredom threshold or
how badly you want a test result).  Both overlapping and nonoverlapping
permutations would then join the diehard tests that are both fixed and
general enough to permit a variable length object to be tested.

This leaves us with diehard_sums, broken as all hell, and some residual
weakness in the diehard monkey tests (opso, oqso, dna) even though I've
recomputed the variances there and fixed the worst of their problems.
In principle one would expect to be able to do the monkey test for
arbitrary numbers of characters, and there is the Marsaglia and Tsang
"gorilla" test -- perhaps I should focus on getting it to work, as well
as adding yet-another-supergenerator, this one one that can be loaded
with a list of generators that return round-robin.  

This generator isn't intended to be used as an example of a good
generator, it is to enable another class of test -- the overlapping seed
test.  The overlapping seed test takes (say) two instances of the
generator of your choice and starts them with seeds that are identical
except for a single bit that may or may not be the least significant
(e.g. 0...00110101 and 0...00100101).  Rands are returned alternating,
and fed to any/all tests that you like.  If the two streams have
significant long term correlation, various tests will fail that don't
fail for the generator for either seed alone.  Testing short term
correlation in the streams post-seed is trickier -- one would need to do
something like reseed the two generators every N returns (for a fairly
small value of N) with bitlevel "near" seeds, selecting new random seeds
for one per cycle and sliding an xor single bit mask down the seed
selected or something like that to see if some particular overlap
results in pathological correlation.

This class of test is important because many MC computations are
performed off of a non-overlapping list of seeds on a cluster computer
of some sort, and it is an open question as to how independent those
runs are.  All the theory of most RNGs guarantees is that for ANY seed
(sometimes only for a carefully selected, not terribly random seed) a
given good generator will produce a good pseudorandom sequence of
numbers; it really says nothing at all about long or short range
correlation or quasiperiodicity in the iterated maps generated by two
distinct but overlapping seeds.

Wa-uh-oh, diehard_operm5 died a horrible death with an Inf error at
10400 psamples.  So perhaps there are few bugs in the fix, perhaps a
failure to reset some variable?  I may have to check this in again in a
second anyway, as I fixed a couple of minor build errors (incorrect
prototype and so on) in the ported code while checking this in and
retesting.

rgb