This fixes the bug I discovered this morning, I think.  What a pain.
The symptom was that ANY test run for 10400 pvalues would fail
(returning either -Nan or -Inf) at -k 1 or -k 2, not just operm5.  I
debugged it with birthdays as it was faster.  Here is what I learned and
what I did as I debugged it.

In the new kstest code, a matrix formed out of the largest deviation
from expected (uniform) rank in a list of psvalues, and that matrix is
taken to more or less the power of the number of psvalues.

In the deviation code I noted two possible (weak) problems:  One is that
the new symmetric code used fmax to select the larger of the left and
right handed deviation from the uniform line while searching for the
maximum deviation.  Previously it looked for an absolute value because
the deviation could easily be negative.  No absolute value was used
before picking the larger fo the left/right deviation, which means that
if the larger one (in magnitude) it could be missed.  So I installed
fabs() calls before the fmax() call which may make a small difference.

I also noted that the uniform line was running from 0 to just under 1,
where the first value was 0.  That was not really correct -- if there
are count pvalues (say 3) and one wants to compare the deviation of the
sorted presumed uniform pvalues from uniform, one would (by hand)
compare the first to 0.25, the second to 0.50, the third to 0.75, not
the first to 0.0, the second to 0.33 and the third to 0.67.  So I added
one to the denominator e.g. i/(count+1) and altered the limits of the
loop so that this now happens -- it starts one step away from 0.0, and
stops one step away from 1.0, so that the pvalues actually correspond
pretty well to the end points.  

However, even before they corresponded sufficiently well that again this
wasn't why the kstest was failing with -Inf for a final pvalue.

So I dug into the actual code.  One thing was immediately the best
candidate for an overflow type error -- the (large!) matrix code
automatically rescales as it runs to avoid just that.  I put in a bunch
of output statements and ran it a whole bunch of times trying to figure
out exactly where and under what circumstances the Nan/Inf was appearing
-- it wasn't just one number but a small range of pvalue counts near
10400, and the failure occurred only if a lot of rescaling occurred of
the matrix as it was multiplying out.  But the code appeared to be
working (right up to where it failed)!  There was nothing >>wrong<< with
the code, it just was overflowing in spite of a rescale step that was
actually happening.

Finally I noticed that in Marsaglia's original rescale conditional he
only tested an element from the middle of the final matrix to the count
power, literally at m/2,m/2 (where m is the size of the square matrix).
If it ALONE exceeded the threshold, the whole matrix was rescaled and
the scale exponent cumulated.  I said to myself, self, what if some
OTHER element in this rather large matrix was getting really big
compared to the particular one in the middle that was being checked?
Maybe this needs to be a double nested loop, one to test EVERY element,
one to rescale if ANY element crosses the threshold...

So I made it so, and it was good, it suddenly worked, no overflow where
before it would overflow every time.  Before I turned off the debugging
statements I also noted that it started rescaling much earlier (and
ended with intermediary matrix elements that were twenty or so orders of
magnitude smaller post rescale), suggesting that some other elements of
the matrix were indeed needing attention before the one being tested.
Go figure.

In any event, I've applied the kstest test to the kstest, and it works
fine out to at least 20,000 psamples after the changes.  I've also run
birthdays and operm5 out to 20000 or more psamples:

 diehard_operm5|   0|   1000000|   25200|0.26310846|  PASSED  

which speaks for itself.  The kstest code is clearly numerically touchy,
and I can't guarantee that it is all quite correct, but it appears to
work long enough for me to get bored waiting for AES to fail in test to
destruction mode (where eventually it will "fail" if only because the
drifting pvalue will drift under its failure threshold as it does the
uniform deviate random walk shuffle).

The good news is that operm5 does indeed seem to work overlapping or not
-- the failure in test to destruct mode didn't LOOK like it was real,
but it was odd and I was worried.  And maybe a small bug in the kstest
is fixed, although it is early days yet and maybe not.