OK, this is a fairly enormously major brutal checkin.  Both dieharder and libdieharder are ALMOST -Wall -pedantic clean.  To get it there I had to learn several things, such as how to get gcc to ignore "unused variables" that are conveniently in a shared include file but aren't really used in all the modules that share it, the fact that the various flavors of C have varying "maximum string size guaranteed to be supported" limits (none of which are really relevant to gcc, but it complains about them anyway), and more.  And of course I had to delete all the cruftish lines of e.g. unused loop variables.  I'm not quite done with cleanup -- I may have gone overboard in a place or two and may need to put some things back or address things that might affect function -- but I want to get this all checked in.
There are two build errors left -- one is in dieharder/rdieharder.c (and
hence is yours, Dirk) and the other is in the skein code (and hence is
yours, David).  David, I also need you to check a fix I made to the
rng_threefish code -- I finally took the time to figure out the dread
"dereferencing type-punned pointer breaks strict aliasing rules"
warning.  I replaced the offending line:

        *((unsigned long int *) state->block) = s;

with

        unsigned long int *blockptr;
...
        blockptr = (unsigned long int*) &state->block;
        *blockptr = s;

That is, I read what you were trying to do as "Set the contents of
state->block, cast to an unsigned long int pointer, equal to unsigned
long int pointer s" which might work but gcc -Wall hated it even before
-pedantic and (from what I've read) can have undesired side effects.  So
I introduced an actual unsigned long int pointer, put the address of
state->block in it, and the set its contents equal to s.  It didn't seem
to break threefish -- I tested the first few returns before and after
the fix with -S 1 and they were the same -- and I'm using threefish
right now in a validate.sh run to make sure that I didn't egregiously
break dieharder with all of the changes.

Changes you should be aware of:

  * To avoid most of the "too long string" errors I went to -std=c95,
which permits strings a page in size (4095 bytes).  That accommodates
the auto-documentation strings in the test headers.  There may be
another way of doing this -- in fact there are probably two or three --
but to alter the dh headers at this point would (marginally) break the
API so it will need to wait for v4, I think.  Apparently gcc is about to
be dressed up with an __attribute__ that will probably enable extra
large or unlimited data strings without complaints which is sensible
enough since it works on them anyway AFAICT.

  * c95 turned off uint translation .  I went through a huge block of
code turning uint into the two words unsigned int before getting
irritated enough to look at the headers where I discovered that yeah,
you can turn on the uint -> unsigned int macro with a suitable define.
So I did.

  * c95 turned of BSD math macros in math.h, including M_PI.  That
seemed really silly, so I turned them back on with a suitable define.  I
didn't turn on the long forms (they only really make sense for long
doubles) but we can do that if we ever need PI to 24 places or whatever
it was.

  * -Wall -pedantic really hates any sort of data that is included in a
source file where it isn't used.  If we were all perfect programmers, I
suppose that we would create enough include files and control where they
were included precisely enough that no source file even included an
include file with a variable it didn't actually use.  Alas, I'm not a
perfect programmer and lots of the data structures used only inside
certain tests or by certain generators are shared via libdieharder.h
with program modules that don't use them.  Adding
__attribute__((unused)) after the definition but before the = sign
basically tells the compiler "Yes, I know, I planned it that way, now
shut the fuck up" and passes them through -pedantic without complaint.

I suppose that the virtue of the check is that it helps prevent
namespace collision, but of course the compiler checks for that anyway
and general local vs global rules seem like they would handle any
accidents that crop up the right way.  If I feel really, really
energetic someday I may go and segregate out the data and either add it
to the sources directly (in a lot of cases that's a good place for it
anyway) or put it in a separate include file per module.  OTOH things
like the dh headers are shared because I DO access content from them in
lots of places and want to be able to get to it from anyplace, so there
will always be some ((unused)) attribute variables in the program.
Printing out the test description string for any given test, or looking
up the default tsamples or psamples, for example, is something any sort
of application that uses the libdieharder library might want to do.

  * As per current GBT recommendations and Dirk's suggestions, all of
the auto-whatever stuff in autogen.sh is now basically a single
autoreconf call.  In fact, it looks like they made autoreconf just
because getting all of the things just right after a major GBT update
is, in fact, the pain in the ass that it has been to me from the
beginning, so this is rather a relief.  I did leave the configure call
in the bottom, so running autogen.sh should still take one from a clean
checkout to make-ready, or of course you can enter autoreconf by hand
and run configure by hand as per usual.  Hopefully this will all make
Debianheads happy...;-)

Things that I have NOT done yet -- this checkin is basically six hours
of work from 9 to 3 am plus another couple of hours today, so I'm
working as fast as I can as it is -- include debugging the endian
problem in the threefish (or was it AES?) code on e.g. a sparc or
powerpc set to the other endianness and dealing with a few real bug
reports that have come in from users already.  I wanted to get the code
clean first as who knows, maybe doing so will help solve the problem?

SO, if you guys could each fix the two remaining problems (or tell me to
play through in spite of the fact that I'm not sure what is being
accomplished and what would break what) then I'll try to move on to the
next step.

   rgb